backend/requirements.txt
# Core Backend
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0
alembic==1.14.0
psycopg2-binary>=2.9.9  # PostgreSQL / Supabase driver

# Vector Database (Local)
chromadb>=0.6.0
sentence-transformers>=3.3.1

# Speech & Audio
openai-whisper==20240930
sounddevice==0.5.1
soundfile==0.12.1
pydub==0.25.1
scipy>=1.15.0
librosa>=0.10.2

# Face & Video Processing
opencv-python>=4.10.0
mediapipe>=0.10.21
Pillow>=11.1.0

# Emotion Detection
fer>=22.5.1
transformers>=4.47.1
torch>=2.5.1
torchaudio>=2.5.1

# Security & Encryption
cryptography>=44.0.0
bcrypt==4.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Email (Local IMAP/POP3)
imapclient==3.0.1
email-validator==2.2.0

# Calendar
icalendar>=6.1.0
recurring-ical-events>=3.5.1

# NLP & Text Processing
spacy>=3.8.3
nltk>=3.9.1
textblob>=0.18.0

# Data Processing
numpy>=2.2.1
pandas>=2.2.3
scikit-learn>=1.6.0

# Task Scheduling
apscheduler>=3.10.4

# Logging & Monitoring
loguru>=0.7.3
rich>=13.9.4

# WebSocket support
websockets>=14.1

# Groq API (cloud LLM)
groq>=0.8.0

# Async HTTP
httpx>=0.27.0
aiofiles>=24.1.0

# Testing
pytest>=8.3.4
pytest-asyncio>=0.25.0

# Voice Cloning - install separately: pip install --no-deps TTS==0.22.0
# TTS==0.22.0
